refactor(tui): move tool output summaries
Move tool argument, tool output, MCP output, image, and diff summary helpers from history.rs into history/tool_output.rs while preserving the existing history re-export paths used by routing and UI callers.

Refs #3308.
